feat: merge List issues with react-hooks example
